requirements.txt
flask
flask-cors
torch
numpy
h5py
pillow
